services/web/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
